klighd.printing: Added exception handling in case printer got removed
PrintOptions could get compromised if the selected printer got removed.
This could for example happen if you are using the interface through a
remote desktop connection. After disconnecting the rdp, no printing is
possible and the workspace needs to be cleaned.
This change catches the generic SWT 'No more Handles' exception when
initializing the cached printer and tries to use the default instead.

Signed-off-by: Nis Wechselberg <nbw@informatik.uni-kiel.de>