klighd: Hopefully fixed edge layout application bug. (KIPRA-1834)
The problem seems to have been that the computation of edge coordinates included the coordinates of ports. However, instead of being based on the new port coordinates as specified by the layout graph, the code used the port coordinates from the view model. Of course, these may not have been updated yet, depending on the order in which the layout results were applied to the view model.

Signed-off-by: Christoph Daniel Schulze <cds@informatik.uni-kiel.de>