klighd: Fixed multiple padding application and dancing edge labels
When starting multiple layout runs in a short time, there was a race
condition that inflated the padding of the nodes.
Instead of adding the insets to the padding in the view model, we now
take the insets into consideration in the configuration store when asked
about the padding. This makes the layout stable again. (KIPRA-1846)

Additionally the "dancing" edge labels have been resovled. This was
caused by nondeterministic order when applying layout data. The change
forces the labels to be handled after the edges, leading to properly
adjusted positions.

Signed-off-by: Nis Wechselberg <nbw@informatik.uni-kiel.de>