incremental: allow dangling elements.
Now instead of failing and throwing an exception whenever a KGraph with
referenced nodes, ports, or edges, which are not in the graph directly
via containment, this tries to still add the elements or display an
error message, but continue doing the best possible out of the faulty
graph. This way the update is way more robust for such graphs.
# Conflicts:
#	plugins/de.cau.cs.kieler.klighd.incremental/src/de/cau/cs/kieler/klighd/incremental/merge/KGraphMerger.java