lsp: switched to lazy tracing of diagram elements.
If the element trace on SModelElements is required for tracing from text
to the diagram or vice versa the traces are now generated lazily only
when they are needed. This speeds up the initial diagram generation
time, especially for very large graphs, where the
PositionConverter#toPosition has an O(n^2) runtime in the model size
because it runs through every character of the model for every model
element.
Tracing from diagram to text is now fast, text to diagram still goes
through the all traces and suffers from above mentioned bad runtime for
very large graphs.

Closes #57.
Closes #90.