lsp: converted package to use plain maven build
This was never meant to be used as an Eclipse plug-in but in plain Java
applications. This transforms the package to a pure Maven build,
configuring all Maven dependencies for this project and other plug-ins
as a bom to be used as a single dependency to build LS with KLighD
diagram support. Also prepares future distribution of fully working
Maven artifacts.