klighd.piccolo.batik: trimmed compile classpath to those jars being actually required by our code, fixes #97
This avoids the 'The package org.w3c.dom is accessible from more than one module: <unnamed>, java.xml' error that is due to fact that classes of the 'org.w3c.dom' package are part of the JDK since Java9 (in module 'java.xml'), and also part of the packaged batik 1.7 distribution.

At runtime classes of the JDK/JRE will take precedence (root class loader is asked first), so their counterparts in the provided jar won't be loaded.