fix handling of edges by incremental update (#171)
* fix handling of edges by incremental update

* incremental: fixed wrong handling of new/old nodes in edges. Added a
more complex test for the incremental update with multiple new elements.

Something is off with the ID generation, possibly causing further issues
(see the test and the FIXME in the KGraphMerger), but solution still
seems to work for the previous issue for now.

* incremental: fixes some issues arising when adding more complex
structures.

This includes nodes that are added together with edges and ports, which
previously could cause issues. Also fixes elements without own IDs being
inserted e.g. in the beginning of lists to not get updated correctly.

---------

Co-authored-by: Niklas Rentz <nre@informatik.uni-kiel.de>