lsp: fix an issue when recursing through non-expanded nodes in the micro layout calculation. (#188)
When generating the SGraph for a KGraph, the SGraph will only be created for any expanded elements and does not recurse further into non-expanded elements. Previously, during the micro layout calculation in the LSP, this still recursed through non-expanded elements, generating IDs for them (resulting in null-IDs) and potentially causing issues and exceptions further down. This fixes the issue by not recursing through the elements either during micro layout, if they are not expanded.
Furthermore, added a check to avoid further NPEs at this place.